refactor: Move some files to core directory
