+++ cmms/core/enums.py
